Moved settings back to being a single file
