+++ backend/api/views.py
@@ -1,4 +1,4 @@
-from backend.settings.constants import INPUT_LIST_LIMIT
+from backend.constants import INPUT_LIST_LIMIT
 from django.contrib.auth.models import User
 from rest_framework import permissions, status, viewsets
 from rest_framework.pagination import PageNumberPagination
